requirements.txt
numpy==1.24.4
pandas==1.5.3
fredapi>=0.5
pyarrow>=10.0,<15.0
python-dotenv>=1.0
matplotlib>=3.7
seaborn>=0.12
scipy>=1.10
statsmodels>=0.14
hmmlearn>=0.3
scikit-learn>=1.3
joblib>=1.3
shap>=0.45
